kok-foong/1/exposure-calculation-agent: consolidate kg_client into utils
